update main
make it where files have a finite size by reading into the data structure and overwriting the file with whatever is part of the data structure. will make it easier to navigate to the right line when we have curses running to push updates, will also make truncating lines not just a front end thing. also, made a struct for a lockable file and run with that instead, will be used for locking only when pushing updates to the file